Move TestHarness to chipyard.harness, make chipyard/harness directory
