+++ fpga/src/main/scala/vcu118/HarnessBinders.scala
@@ -9,8 +9,8 @@ import freechips.rocketchip.tilelink.{TLBundle}
 import sifive.blocks.devices.uart.{HasPeripheryUARTModuleImp, UARTPortIO}
 import sifive.blocks.devices.spi.{HasPeripherySPI, SPIPortIO}
 
-import chipyard.{HasHarnessSignalReferences, CanHaveMasterTLMemPort}
-import chipyard.harness.{OverrideHarnessBinder}
+import chipyard.{CanHaveMasterTLMemPort}
+import chipyard.harness.{HasHarnessSignalReferences, OverrideHarnessBinder}
 
 /*** UART ***/
 class WithUART extends OverrideHarnessBinder({
